+++ extensions/deepinfra/media-understanding-provider.ts
@@ -11,7 +11,7 @@ import {
   DEEPINFRA_STT_FALLBACK_MODELS,
   DEEPINFRA_VLM_FALLBACK_MODELS,
 } from "./media-models.js";
-import type { DeepInfraSurfaceModel } from "./provider-models.js";
+import type { DeepInfraSurfaceModel } from "./media-models.js";
 
 function resolveDefault(
   surfaceModels: readonly DeepInfraSurfaceModel[] | undefined,
