+++ extensions/deepinfra/onboard.test.ts
@@ -6,12 +6,9 @@ import {
 } from "openclaw/plugin-sdk/provider-onboard";
 import { captureEnv } from "openclaw/plugin-sdk/test-env";
 import { afterEach, describe, expect, it, vi } from "vitest";
+import { DEEPINFRA_BASE_URL } from "./media-models.js";
 import { applyDeepInfraConfig } from "./onboard.js";
-import {
-  DEEPINFRA_BASE_URL,
-  DEEPINFRA_DEFAULT_MODEL_REF,
-  DEEPINFRA_MODEL_CATALOG,
-} from "./provider-models.js";
+import { DEEPINFRA_DEFAULT_MODEL_REF, DEEPINFRA_MODEL_CATALOG } from "./provider-models.js";
 
 const { resolveEnvApiKey } = providerAuth;
