+++ extensions/deepinfra/provider-catalog.ts
@@ -5,8 +5,8 @@ import {
   type ProviderCatalogResult,
 } from "openclaw/plugin-sdk/provider-catalog-shared";
 import type { ModelProviderConfig } from "openclaw/plugin-sdk/provider-model-shared";
+import { DEEPINFRA_BASE_URL } from "./media-models.js";
 import {
-  DEEPINFRA_BASE_URL,
   DEEPINFRA_MODEL_CATALOG,
   buildDeepInfraModelDefinition,
   discoverDeepInfraModels,
