+++ extensions/deepinfra/speech-provider.ts
@@ -11,7 +11,7 @@ import {
   DEFAULT_DEEPINFRA_TTS_VOICE,
   normalizeDeepInfraModelRef,
 } from "./media-models.js";
-import type { DeepInfraSurfaceModel } from "./provider-models.js";
+import type { DeepInfraSurfaceModel } from "./media-models.js";
 
 const DEEPINFRA_TTS_RESPONSE_FORMATS = ["mp3", "opus", "flac", "wav", "pcm"] as const;
