+++ extensions/deepinfra/surface-model-catalogs.ts
@@ -8,7 +8,8 @@ import type {
   VideoGenerationProviderCapabilities,
 } from "openclaw/plugin-sdk/video-generation";
 import { DEEPINFRA_VIDEO_ASPECT_RATIOS, DEEPINFRA_VIDEO_DURATIONS } from "./media-models.js";
-import { discoverDeepInfraSurfaces, type DeepInfraSurfaceModel } from "./provider-models.js";
+import type { DeepInfraSurfaceModel } from "./media-models.js";
+import { discoverDeepInfraSurfaces } from "./provider-models.js";
 
 const PROVIDER_ID = "deepinfra";
