+++ src/plugin-sdk/test-helpers/provider-discovery-contract.ts
@@ -139,9 +139,9 @@ function installDiscoveryHooks(state: DiscoveryState, options: DiscoveryContract
   beforeAll(async () => {
     vi.resetModules();
     vi.doMock("openclaw/plugin-sdk/provider-auth", async (importOriginal) => {
-      const { findNormalizedProviderValue, resolveAuthProfileOrder } =
-        await importOriginal<typeof import("../provider-auth.js")>();
+      const actual = await importOriginal<typeof import("../provider-auth.js")>();
       return {
+        ...actual,
         DEFAULT_COPILOT_API_BASE_URL: "https://api.individual.githubcopilot.com",
         MINIMAX_OAUTH_MARKER: "minimax-oauth",
         applyAuthProfileConfig: (config: OpenClawConfig) => config,
@@ -163,7 +163,6 @@ function installDiscoveryHooks(state: DiscoveryState, options: DiscoveryContract
         coerceSecretRef: asNullableRecord,
         ensureApiKeyFromOptionEnvOrPrompt: vi.fn(),
         ensureAuthProfileStore: ensureAuthProfileStoreMock,
-        findNormalizedProviderValue,
         listProfilesForProvider: listProfilesForProviderMock,
         normalizeApiKeyInput: (value: unknown) => (typeof value === "string" ? value.trim() : ""),
         normalizeGithubCopilotDomain: (raw: unknown) => {
@@ -173,7 +172,6 @@ function installDiscoveryHooks(state: DiscoveryState, options: DiscoveryContract
             : "github.com";
         },
         normalizeOptionalSecretInput: normalizeOptionalString,
-        resolveAuthProfileOrder,
         resolveNonEnvSecretRefApiKeyMarker: (source: unknown) =>
           typeof source === "string" ? source : "",
         upsertAuthProfile: vi.fn(),
